feat: remove ink recognition support in the fdroid flavor
